Move functions to misc/statistics
Remove statistics-related functions from Had_DCAN.
Add function to generate Covariance and Average matrices from STL-10.
