+++ spyrit/learning/model_Had_DCAN.py
@@ -8,25 +8,25 @@
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
-import torch.optim as optim
+#import torch.optim as optim
 import numpy as np
-import torchvision
-from torchvision import datasets, models, transforms
-import matplotlib.pyplot as plt
-import time
-import os
-import copy
+#import torchvision
+#from torchvision import datasets, models, transforms
+#import matplotlib.pyplot as plt
+#import time
+#import os
+#import copy
 from fht import *
 from ..misc.pattern_choice import Hadamard, matrix2conv, split
 from collections import OrderedDict
-import cv2
+#import cv2
 from scipy.stats import rankdata
-#from ..misc.disp import *
-from itertools import cycle;
-from pathlib import Path
+#from itertools import cycle;
+#from pathlib import Path
 
 from ..misc.disp import *
-import spyrit.misc.walsh_hadamard as wh
+#import spyrit.misc.walsh_hadamard as wh
+from spyrit.misc.statistics import *
 import math
 
 
@@ -36,29 +36,6 @@ import math
 # Helps determining the statistical best 
 # Hadamard patterns for a given image size
 # 
-
-def optim_had(dataloader, root):
-    """ Computes image that ranks the hadamard coefficients
-    """
-    inputs, classes = next(iter(dataloader))
-    inputs = inputs.cpu().detach().numpy();
-    (batch_size, channels, nx, ny) = inputs.shape;
-
-    tot_num = len(dataloader)*batch_size;
-    Cumulated_had = np.zeros((nx, ny));
-    # Iterate over data.
-    for inputs,labels in dataloader:
-        inputs = inputs.cpu().detach().numpy();
-        for i in range(inputs.shape[0]):
-            img = inputs[i,0,:,:];
-            h_img = np.abs(fht2(img))/tot_num;
-            Cumulated_had += h_img;
-    
-    Cumulated_had = Cumulated_had / np.max(Cumulated_had) * 255
-    np.save(root+'{}x{}'.format(nx,ny)+'.npy', Cumulated_had)
-    np.savetxt(root+'{}x{}'.format(nx,ny)+'.txt', Cumulated_had)
-    cv2.imwrite(root+'{}x{}'.format(nx,ny)+'.png', Cumulated_had)
-    return Cumulated_had 
 
 def hadamard_opt_spc(M ,root, nx, ny):
     msk = np.ones((nx,ny))
@@ -71,194 +48,6 @@ def hadamard_opt_spc(M ,root, nx, ny):
     return conv
 
 
-def abs_walsh(dataloader, device):
-    
-    # Estimate tot_num
-    inputs, classes = next(iter(dataloader))
-    #inputs = inputs.cpu().detach().numpy();
-    (batch_size, channels, nx, ny) = inputs.shape;
-    tot_num = len(dataloader)*batch_size;
-    
-    # Init
-    n = 0
-    output = torch.zeros((nx,ny),dtype=torch.float32)
-    H = wh.walsh_matrix(nx).astype(np.float32, copy=False)
-    
-    # Send to device (e.g., cuda)
-    output = output.to(device)
-    H = torch.from_numpy(H).to(device)
-    
-    # Accumulate over all images in dataset
-    for inputs,_ in dataloader:
-        inputs = inputs.to(device);
-        n = n + inputs.shape[0]
-        trans = wh.walsh2_torch(inputs,H);
-        trans = torch.abs(trans)
-        output = output.add(torch.sum(trans,0))
-        print(f'Abs:  {n} / (less than) {tot_num} images', end='\n')
-    print('', end='\n')
-    
-    #-- Normalize
-    output = output/n;
-    output = torch.squeeze(output)
-    
-    return output
-
-def stat_walsh(dataloader, device, root):
-    
-    # Get dimensions and estimate total number of images in the dataset
-    inputs, classes = next(iter(dataloader))
-    (b, c, nx, ny) = inputs.shape;
-    tot_num = len(dataloader)*b;
-    
-    # 1. Mean
-    
-    # Init
-    n = 0
-    mean = torch.zeros((nx,ny), dtype=torch.float32)
-    H = wh.walsh_matrix(nx).astype(np.float32, copy=False)
-    
-    # Send to device (e.g., cuda)
-    mean = mean.to(device)
-    H = torch.from_numpy(H).to(device)
-    
-    # Accumulate sum over all images in dataset
-    for inputs,_ in dataloader:
-        inputs = inputs.to(device);
-        trans = wh.walsh2_torch(inputs,H)
-        mean = mean.add(torch.sum(trans,0))
-        # print
-        n = n + inputs.shape[0]
-        print(f'Mean:  {n} / (less than) {tot_num} images', end='\n')
-    print('', end='\n')
-    
-    # Normalize
-    mean = mean/n;
-    mean = torch.squeeze(mean)
-    #torch.save(mean, root+'Average_{}x{}'.format(nx,ny)+'.pth')
-    np.save(root / Path('Average_{}x{}'.format(nx,ny)+'.npy'), mean.cpu().detach().numpy())
-    
-    # 2. Covariance
-    
-    # Init
-    n = 0
-    cov = torch.zeros((nx*ny,nx*ny), dtype=torch.float32)
-    cov = cov.to(device)
-    
-    # Accumulate (im - mu)*(im - mu)^T over all images in dataset
-    for inputs,_ in dataloader:
-        inputs = inputs.to(device);
-        trans = wh.walsh2_torch(inputs,H)
-        trans = trans - mean.repeat(inputs.shape[0],1,1,1)
-        trans = trans.view(inputs.shape[0], nx*ny, 1)
-        cov = torch.addbmm(cov, trans, trans.view(inputs.shape[0], 1, nx*ny))
-        # print
-        n += inputs.shape[0]
-        print(f'Cov:  {n} / (less than) {tot_num} images', end='\n')
-    print('', end='\n')
-    
-    # Normalize
-    cov = cov/(n-1);
-    #torch.save(cov, root+'Cov_{}x{}'.format(nx,ny)+'.pth') # todo?
-    np.save(root / Path('Cov_{}x{}'.format(nx,ny)+'.npy'), cov.cpu().detach().numpy())
-    
-    return mean, cov
-
-def stat_walsh_np(dataloader, root):
-    """ 
-        Computes Mean Hadamard Image over the whole dataset + 
-        Covariance Matrix Amongst the coefficients
-    """
-    inputs, classes = next(iter(dataloader))
-    inputs = inputs.cpu().detach().numpy();
-    (batch_size, channels, nx, ny) = inputs.shape;
-    tot_num = len(dataloader)*batch_size;
-    
-    H1d = wh.walsh_ordered(nx)
-    
-     # Abs matrix
-    Mean_had = abs_walsh_ordered(dataloader, H1d, tot_num)
-    print("Saving abs")
-    np.save(root / Path('Abs_{}x{}'.format(nx,ny)+'.npy'), Mean_had)
-
-    # Mean matrix
-    #-- Accumulate over all images in dataset
-    n = 0
-    Mean_had = np.zeros((nx, ny));
-    for inputs,_ in dataloader:
-        inputs = inputs.cpu().detach().numpy();
-        for i in range(inputs.shape[0]):
-            img = inputs[i,0,:,:];
-            h_img = wh.walsh_ordered2(img,H1d);
-            Mean_had += h_img;
-            n = n+1
-        print(f'Mean:  {n} / (less than) {tot_num} images', end='\r')
-    print('', end='\n')
-    
-    #-- Normalize & save
-    Mean_had = Mean_had/n;
-    print("Saving mean")
-    np.save(root / Path('Mean_{}x{}'.format(nx,ny)+'.npy'), Mean_had)
-    
-    # Covariance matrix    
-    n = 0
-    Cov_had = np.zeros((nx*ny, nx*ny));
-    for inputs,_ in dataloader:
-        inputs = inputs.cpu().detach().numpy();
-        for i in range(inputs.shape[0]):
-            img = inputs[i,0,:,:];
-            h_img = walsh_ordered2(img, H1d);
-            Norm_Variable = np.reshape(h_img-Mean_had, (nx*ny,1));
-            Cov_had += Norm_Variable*np.transpose(Norm_Variable);
-            n = n+1
-        print(f'Covariance:  {n} / (less than) {tot_num} images', end='\r')     
-    print()
-    
-    #-- Normalize & save
-    Cov_had = Cov_had/(n-1);  
-    np.save(root / Path('Cov_{}x{}'.format(nx,ny)+'.npy'), Cov_had)
-
-
-
-def Stat_had(dataloader, root):
-    """ 
-        Computes Mean Hadamard Image over the whole dataset + 
-        Covariance Matrix Amongst the coefficients
-    """
-
-    inputs, classes = next(iter(dataloader))
-    inputs = inputs.cpu().detach().numpy();
-    (batch_size, channels, nx, ny) = inputs.shape;
-    tot_num = len(dataloader)*batch_size;
-
-    Mean_had = np.zeros((nx, ny));
-    for inputs,labels in dataloader:
-        inputs = inputs.cpu().detach().numpy();
-        for i in range(inputs.shape[0]):
-            img = inputs[i,0,:,:];
-            h_img = fht2(img);
-            Mean_had += h_img;
-    Mean_had = Mean_had/tot_num;
-
-    Cov_had = np.zeros((nx*ny, nx*ny));
-    for inputs,labels in dataloader:
-        inputs = inputs.cpu().detach().numpy();
-        for i in range(inputs.shape[0]):
-            img = inputs[i,0,:,:];
-            h_img = fht2(img);
-            Norm_Variable = np.reshape(h_img-Mean_had, (nx*ny,1));
-            Cov_had += Norm_Variable*np.transpose(Norm_Variable);
-    Cov_had = Cov_had/(tot_num-1);
-
-    np.save(root+'Cov_{}x{}'.format(nx,ny)+'.npy', Cov_had)
-    np.savetxt(root+'Cov_{}x{}'.format(nx,ny)+'.txt', Cov_had)
-    
-    np.save(root+'Average_{}x{}'.format(nx,ny)+'.npy', Mean_had)
-    np.savetxt(root+'Average_{}x{}'.format(nx,ny)+'.txt', Mean_had)
-    cv2.imwrite(root+'Average_{}x{}'.format(nx,ny)+'.png', Mean_had) #Needs conversion to Uint8!
-    return Mean_had, Cov_had 
-
-
 def img2mask(Value_map, M):
     (nx, ny) = Value_map.shape;
     msk = np.ones((nx, ny));
@@ -266,15 +55,6 @@ def img2mask(Value_map, M):
     msk[np.absolute(ranked_data)>M]=0;
     return msk
 
-def Cov2Var(Cov):
-    """
-    Extracts Variance Matrix from Covarience Matrix
-    """
-    (Nx, Ny) = Cov.shape;
-    diag_index = np.diag_indices(Nx);
-    Var = Cov[diag_index];
-    Var = np.reshape(Var, (int(np.sqrt(Nx)),int(np.sqrt(Nx))) );
-    return Var
 
 def Permutation_Matrix_root(root):
     """
